Fix NPE when checking for already promoted version
Prior to this commit, a failed staging request was followed by an
automated check for already promoted versions. In some cases, this check
fails with a client error - this case is alreaydy handled. If the check
gets a response from the server with a missing "statuses" response
field, this would fail with a NPE.

This commit ensures that we defensively check for the presence of this
attribute before reading it. If it's not present, the version has not
been promoted already and we can let the original exception bubble up to
the application.

Fixes gh-9