Fix hard coded reference to Spring Boot artifact
Prior to this commit, the `SonatypeService` would use a hard coded
reference to a Spring Boot artifact when checking if the released
artifacts were already published to the release nexus repository.

This commit selects a JAR artifact from the build info and uses its
coordinates to check if it's been published to the repository.

Fixes gh-11