Configure sonatype task with staging profile name
Prior to this commit, developers would need to find the
staging_profile_id information in their Nexus instance for staging
artifacts with the right profile.

This commit adds a new configuration property,
`sonatype.staging_profile` which takes the name of the staging profile
and resolves the id automatically. The former configuration key is now
deprecated and scheduled for removal.

Fixes gh-13